+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/service/projectgenerator/factory/ProjectGeneratorFactoryPhase.java
@@ -90,7 +90,8 @@ public class ProjectGeneratorFactoryPhase implements ProjectGeneratorFactory {
               phase,
               null,
               startDate,
-              saleOrderLine.getQty());
+              saleOrderLine.getQty(),
+              saleOrderLine);
         }
       }
     }